+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add ``mobile-sort-placeholder`` prop to table
 * Add ``contextmenu`` event to table (thanks @gps-dev)
 * Add ``show-detail-icon`` prop to table and toggle row detail programmatically (thanks @omer2500)
+* Add ``use-html5-validation`` prop to form components and ``defaultUseHtml5Validation` constructor options (thanks @lukio3)
 
 ### Fixes